requirements.txt
pillow
numpy
tqdm

# Qwen2.5-VL: nécessite une version récente de transformers.
# (Sinon AutoProcessor peut planter dans les modules "video_processing_auto".)
transformers>=4.48.0
accelerate>=0.30.0
datasets>=2.18.0

# Utilitaire officiel Qwen pour préparer les entrées vision (image/video placeholders)
# NOTE: qwen-vl-utils nécessite torchvision (dépendance non déclarée)
qwen-vl-utils>=0.0.8

# GPT-4o eval via OpenAI API
openai>=1.0.0

# Laissez torch/torchvision être géré côté serveur CUDA si besoin (pip/conda).
# Si vous voulez pip ici, décommentez les lignes ci-dessous avec la wheel CUDA appropriée.
# torch
# torchvision
